test: refactor AmendTask import paths
